requirements.txt
asgiref==3.4.1
blis==0.7.5
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
cymem==2.0.6
Cython==0.29.24
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.6
django-admin-interface==0.16.3
django-colorfield==0.4.2
django-database-view==0.3.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-font-awesome==4.6.1
django-heroku==0.3.1
django-icons==4.0.0
django-import-export==2.5.0
et-xmlfile==1.1.0
fontawesome-free==5.15.4
gensim==4.1.2
gunicorn==20.1.0
idna==3.3
inexactsearch==1.0.2
Jinja2==3.0.3
joblib==1.1.0
langcodes==3.3.0
MarkupPy==1.14
MarkupSafe==2.0.1
murmurhash==1.0.6
mysqlclient==2.0.3
nltk==3.7
numpy==1.21.2
ODENet==0.0.5
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
pandas==1.3.3
pathy==0.6.1
Pillow==8.4.0
preshed==3.0.6
psycopg2==2.9.1
py-openthesaurus==1.0.5
pydantic==1.8.2
pygermanet==1.0.2
pymongo==4.0.1
pyparsing==3.0.6
pyspellchecker==0.6.3
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.11.10
requests==2.26.0
scikit-learn==1.0.2
scipy==1.8.0
silpa-common==0.3
six==1.16.0
smart-open==5.2.1
soundex==1.1.3
spacy==3.2.0
spacy-legacy==3.0.8
spacy-loggers==1.0.1
spellchecker==0.4
sqlparse==0.4.1
srsly==2.4.2
tablib==3.0.0
thinc==8.0.13
threadpoolctl==3.1.0
tomli==2.0.1
tqdm==4.62.3
typer==0.4.0
typing_extensions==4.0.0
urllib3==1.26.7
wasabi==0.8.2
whitenoise==5.3.0
wn==0.9.1
xlrd==2.0.1
xlwt==1.3.0
